Change the PaymentRequestUpdateEvent model to better match FetchEvent
Previously there was a bug in the event model where if a page didn't
listen for the update event and call updatePaymentRequest then the user
interface would be permanently disabled. This change modifies the model
to mirror the way that the FetchEvent works in a ServiceWorker. If a
page wishes to carry out an update then it calls updateWith and provides
a promise that will settle when the page is ready to proceed. If it is
resolved with a PaymentDetails dictionary then the payment request will
be updated with the supplied information. If it is rejected then no
update occurs and the request can continue. If the page does not listen
for the event or does not call updateWith then the request can continue.